python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
sentence-transformers>=2.2.0
torch>=2.6.0
numpy==1.26.4
scikit-learn==1.5.2
umap-learn==0.5.6
mistralai>=1.0.0
python-dotenv==1.0.1
scipy>=1.11.0
cohere>=5.0.0
# Optional: SONAR (Meta) - requires fairseq2
# Install separately: pip install sonar-space
